--- src/sssom_pydantic/web/__init__.py
@@ -1,9 +1,0 @@
-"""Implementation of a semantic mappings API."""
-
-from .impl import get_app
-from .router import router
-
-__all__ = [
-    "get_app",
-    "router",
-]
